Update all tests to use the right constructor
In 395d1ecdab441b1ad7b728cadc69f4dfb49dcd6f, the CodecNew constructor
was reverted to only take two booleans, and a new CodecNewFast with an
option for fast parsing was added.

That caused all tests to fail with:

    ..._test.go:16:48: too many arguments in call to ais.CodecNew
        have (bool, bool, bool)
        want (bool, bool)

Update them to use the right constructor depending on their needs.